TranscriptScreen: increase mTotalRows when resizing if needed
Currently, when you attempt to increase the number of screen rows in a
TranscriptScreen beyond the value of mTotalRows supplied when the
TranscriptScreen was initially created, the backing store allocated in
init() will not be large enough to hold the contents of the screen.
This results in strange behavior, such as duplicate lines or lines
overwriting each other in full-screen applications.

Fix this by bumping mTotalRows as needed when resizing.