Draw combining characters at the ends of runs correctly
At the moment, TranscriptScreen's drawTextRun() tracks the screen column
corresponding to the character it's working on at the moment by updating
the column after processing each character; this results in combining
characters being considered part of the column after the one they should
apply to.

For most purposes, this does not matter -- Android's own text drawing
routines don't care about our column count, so the combining characters
will be drawn with the correct base character.  However, there are
corner cases where this causes problems:

* Combining characters in the last screen column will be ignored
  entirely (after processing the base character, the column count
  exceeds the number of columns, so processing stops)
* If the last column of a text run (sequence with the same style, or a
  selection range) has combining characters, they will be erroneously
  drawn outside the run (the run ends after the base character, so
  the combining characters aren't passed to the text drawing routine as
  part of the run)

Instead, update a next-column counter after processing each character,
and use it to set the column count once we hit the next spacing
character.  This requires us to add array bounds checking to the loop,
as it's now possible to fall off the end of the array looking for the
next column; that has the side effect of papering over many possible
crash bugs involving line buffer corruption.