UnicodeTranscript: add fast path for getLine() on an entire FullUnicodeLine
getLine() is called frequently during screen drawing and other
operations, usually with a request for the contents of an entire line.
We have a fast path for this case with basic lines; something similar
works for FullUnicodeLines, so let's do that too.

Also, update the documentation to reflect the fact that the array
returned from getLine() will not be null-terminated if the array size
exactly matches the length of the requested text.