Fix selecting/copying text runs bookended by East Asian wide characters
Text runs beginning or ending with East Asian wide characters can
currently behave strangely when selected:

* An East Asian wide character immediately before the displayed
  selection is sometimes copied along with the highlighted text, despite
  not being highlighted;
* An East Asian wide character at the end of the displayed selection is
  sometimes omitted from the copied text, despite being highlighted.

Both of these problems have the same root cause: EmulatorView's
reporting of selection bounds does not take the width of the underlying
character into account, even though the drawing of the selection
highlight does.  In other words, it's possible for EmulatorView to ask
for half of an East Asian wide character to be selected, but the
highlight won't reflect this (either the entire character is
highlighted, or none of it is).

Preventing EmulatorView from reporting selection bounds like this would
require making the text selection routines aware of the underlying
characters they operate on (as opposed to just working from screen
positions).  Instead:

* Have UnicodeTranscript include half-selected East Asian wide characters
  at the end of a selection in copied text.
* Ensure that a half-selected East Asian wide character at the beginning
  of a selection is highlighted.
* Adjust UnicodeTranscript.getLineColor() to ensure that the number of
  columns of color information returned in a partial line always matches
  the number of columns in the text that getLine() would return.
* Ensure that TranscriptScreen.internalGetTranscriptText() doesn't
  truncate the returned text when there are half-selected East Asian
  wide characters at both the beginning and end of the selection range.
  This requires us to avoid bounds checking based on selection width
  (the naive computation based on screen positions would be off by two,
  omitting the last character, and we lack the information here to do
  better); using a try-catch clause to replace the bounds check is
  likely to paper over some crash bugs involving line buffer corruption.

This ensures a consistent user experience: half-selected East Asian wide
characters are always highlighted and always included in the copied
text, whether at the beginning or end of the selection.