Term: ignore invalid attempts to use private intents
The main Term activity implements a set of private intents for use by
the remote interfaces.  These are only supposed to be used via the
TermInternal activity alias, which is private to TEA, but it's possible
for another application to invoke these on the main Term activity by
explicitly specifying a component name (as in #374).  It should not be
possible to do more than crash TEA via these intents, but let's prevent
outside applications from using them anyway by ignoring requests which
don't come via the TermInternal alias.